[XProto] Replace usages of XID and ::Window with x11::Window
https://chromium-review.googlesource.com/c/chromium/src/+/2249389
(cherry picked from commit 97f5b59e066cecb5e93e6e31a6c276bb6681645e)

+++ shell/browser/ui/x/window_state_watcher.cc
@@ -57,7 +57,8 @@ void WindowStateWatcher::DidProcessXEvent(x11::Event* x11_event) {
 bool WindowStateWatcher::IsWindowStateEvent(x11::Event* x11_event) const {
   XEvent* xev = &x11_event->xlib_event();
   return (static_cast<x11::Atom>(xev->xproperty.atom) == window_state_atom_ &&
-          xev->type == PropertyNotify && xev->xproperty.window == widget_);
+          xev->type == PropertyNotify &&
+          xev->xproperty.window == static_cast<uint32_t>(widget_));
 }
 
 }  // namespace electron
